Used the sensors to detect steps for the user by measuring the acceleration of the mobile. To detect a step, I could use an algorithm. The algorithm steps: Step one : I could Compute the magnitude of the acceleration vector. This the calculation used :
magnitude = sqrt(event.x * event.x + event.y * event.y + event.z * event.z)

Then to move to the second step get the delta this is the difference between the current magnitude and the previous magnitude.
I need to check If the delta is greater than a certain threshold, and the current magnitude is above a certain threshold,consequently count a step.
Then decide  If the current magnitude is below a certain threshold no count the step.